refactor: waku/router/server files
