+++ packages/waku/src/client.ts
@@ -207,6 +207,10 @@ export const Slot = ({
 
 export const Children = () => use(ChildrenContext);
 
+/**
+ * ServerRoot for SSR
+ * This is not a public API.
+ */
 export const ServerRoot = ({
   elements,
   children,
